docs: purge legacy execution-mode vocabulary from quickstart, README, docs, examples
Quickstart/onboarding no longer force mock/offline or mention algorithm/offline;
README/docs/reference describe only the algorithm+offline model (auto=cloud smart
default, grid/random=local, smart=cloud-only; results sync unless offline). Rename
the execution-modes example to optimization-routing; drop the local-basic and
local-privacy demos. BE /hybrid API-contract docs retained (backend names).
