+++ docs/examples/QUICK_REFERENCE.md
@@ -30,12 +30,11 @@ def summarize(text: str) -> str:
     return f"Summary | model={cfg['model']}"  # call your LLM here
 ```
 
-## Execution model
-- `algorithm="auto"` (default) - cloud-first optimizer decisions, local trials, local fallback on connectivity failures.
-- `algorithm="grid"` / `"random"` - explicit local optimizers with no cloud optimizer round trip.
+## Optimization routing
+- Omit routing settings for the default smart optimizer and portal sync.
+- `algorithm="grid"` / `"random"` runs local search and still syncs results when authenticated.
 - Smart algorithms such as `bayesian`, `tpe`, `optuna_tpe`, `nsga2`, and `cmaes` are cloud-only and hard-error if unavailable.
-- `offline=True`, `TRAIGENT_OFFLINE=1`, or legacy `TRAIGENT_OFFLINE_MODE=1` - zero Traigent backend egress.
-- `TRAIGENT_REQUIRE_CLOUD=1` - disables the automatic local fallback for `algorithm="auto"`.
+- `offline=True` is the zero Traigent backend egress path and does not sync to the portal.
 - Result provenance: `result.source` is `cloud_brain`, `local_fallback`, `explicit_local`, or `offline`.
 - Mock: call `traigent.testing.enable_mock_mode_for_quickstart()` in local tutorial code.
