+++ examples/gallery/readme.html
@@ -40,7 +40,7 @@
             <div class="nav-title">Learn Traigent</div>
             <a href="../../docs/user-guide/README.md" class="nav-item">Core Concepts</a>
             <a href="../../docs/user-guide/configuration-spaces.md" class="nav-item">Configuration Management</a>
-            <a href="../../docs/guides/execution-modes.md" class="nav-item">Execution Modes</a>
+            <a href="../../docs/guides/execution-modes.md" class="nav-item">Optimization Routing</a>
             <a href="../../docs/guides/evaluation.md" class="nav-item">Results & Analysis</a>
         </div>
